requirements.txt
matplotlib
pandas
scikit-learn
pillow
numpy
tqdm
faiss-gpu
opencv-python
opencv-contrib-python
transformers
datasets
trimesh
pyglet